backend/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
python-jose==3.3.0